database/requirements.txt
# Database Integration Requirements
# Install with: pip install -r requirements.txt

# Database connector
mysql-connector-python==8.2.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP requests
requests==2.31.0

# RASA SDK (for custom actions)
rasa-sdk==3.6.0

# Utilities
python-dotenv==1.0.0
